% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgme_vary.R
\name{[[.dgme_vary}
\alias{[[.dgme_vary}
\title{Extract an alternative economy by index}
\usage{
\method{[[}{dgme_vary}(x, i, ...)
}
\arguments{
\item{x}{A \code{dgme_vary} object.}

\item{i}{Integer index of the alternative economy to extract.}

\item{...}{Further arguments (currently unused).}
}
\value{
A \code{dgme_params} object.
}
\description{
Extract an alternative economy by index
}
